Lexer Code Cleanup & Improvements
The isEOF method  was updated, because the "state" variable it used is not needed anymore since I've migrated to a better Error management system. (The rest of the interpreting process is automatically aborted when errors have been thrown ealier)

I Also deleted the method "sanityCheck", which used to check if inChar and inStr were both active at the same time. It was useless, since one of the first actions of the lexing loop is checking if it's in a char/string and if it is, it just adds the characters blindly without checking if they're a ' or a ".