Lexer bug  solved. Typos solving and Indentation improvements.
I Had a dumb typo in the Data class that caused ints to be treated as floats, probably due to some late night coding. Same in the error class, where I had empty if statement (I Added a ';' at the end of the 'if' accidentally, that caused the if to be ignored and the statement to be executed anyway.  It solved a warning too !

Also, I'm an idiot. I used != to invert a boolean value, (lines 52/54 in Lexer.cpp) but it's a comparison operator.. This revealed a pretty big bug in the way strings were treated in my lexer so I reworked that bit of the code and I think it's Ok now !

I also modified a bit the piece of code between the line 118 and 132 of Data.cpp, about char identification. It will now throw warnings if a char size is different from 3 (Lower : empty char -> defaults to ' ', Higher : Char is too big : Only takes the first char)

I Also removed a few useless empty lines.