Added comments support and more error detection
I Added comments ! Use them with the C-Like syntax : //
Everything after the double slash will be ignored until the end of the line.

While testing this, I found a huge bug : if you left a tag unclosed, the interpreter would crash with "Access Violation" :(. I added a check in the MakeTree method (Parser.cpp @ line 109) to check for unclosed tags, and throw an error if there's one. (Because there might be errors while accessing the tree later, it's too dangerous to search it.)

I updated the error.h dictionary,and added a "[Message:]" text before the custom error messages for clarity.